+++ NEWS.md
@@ -4,12 +4,15 @@ taxlist 0.1.5
 ### New Features
 
 * A **CITATION** file is included in the installation.
+* New method `replace_view`.
 
 ### Improvements
 
 * Function `accepted_name` retrieves also information on `Level` (taxonomic rank) and traits (optional in argument `get_traits`).
 * Function `summary` for single taxon is displaying the name of the parent taxon (accepted name) and optional a string for the taxon view.
 * Function `backup_object` prints a message in the console.
+* Related functions will join documentation files.
+* Data set `Easplist` adapted to new state of database **SWEA-Dataveg**.
 
 ### Bugs Fixes
 * Function `load_last` was not working for single files with suffix