+++ NEWS.md
@@ -8,6 +8,7 @@ taxlist 0.1.5
 * New method `print_name` for formatting taxon names to italic style.
 * New method `update_name`, for updating information in slot `taxonNames`.
 * New method `synonyms` retrieving synonyms for indicated concepts.
+* New method `delete_name` for deleting synonyms in `taxlist` objects.
 
 ### Improvements